move platform_sample test cases to separate file for reusability
