+++ pants-plugins/uses_services/mongo_rules_test.py
@@ -18,6 +18,7 @@ import pytest
 from pants.engine.internals.scheduler import ExecutionError
 from pants.testutil.rule_runner import QueryRule, RuleRunner
 
+from .data_fixtures import platform, platform_samples
 from .exceptions import ServiceMissingError
 from .mongo_rules import (
     MongoIsRunning,
@@ -59,33 +60,6 @@ def run_mongo_is_running(
     return result
 
 
-def platform(
-    arch="",
-    os="",
-    distro="",
-    distro_name="",
-    distro_codename="",
-    distro_like="",
-    distro_major_version="",
-    distro_version="",
-    mac_release="",
-    win_release="",
-) -> Platform:
-    """Create a Platform with all values defaulted to the empty string."""
-    return Platform(
-        arch=arch,
-        os=os,
-        distro=distro,
-        distro_name=distro_name,
-        distro_codename=distro_codename,
-        distro_like=distro_like,
-        distro_major_version=distro_major_version,
-        distro_version=distro_version,
-        mac_release=mac_release,
-        win_release=win_release,
-    )
-
-
 # Warning this requires that mongo be running
 def test_mongo_is_running(rule_runner: RuleRunner) -> None:
     request = UsesMongoRequest()
@@ -96,70 +70,7 @@ def test_mongo_is_running(rule_runner: RuleRunner) -> None:
     assert is_running
 
 
-@pytest.mark.parametrize(
-    "mock_platform",
-    (
-        platform(),  # empty
-        # platform(
-        #    arch="x86_64",
-        #    os="Linux",
-        #    distro="",
-        #    distro_name="",
-        #    distro_codename="",
-        #    distro_like="",
-        #    distro_major_version="",
-        #    distro_version="",
-        # ),
-        platform(
-            arch="x86_64",
-            os="Linux",
-            distro="centos",
-            distro_name="Centos Linux",
-            distro_codename="Core",
-            distro_like="rhel fedora",
-            distro_major_version="7",
-            distro_version="7",
-        ),
-        platform(
-            arch="x86_64",
-            os="Linux",
-            distro="ubuntu",
-            distro_name="Ubuntu",
-            distro_codename="xenial",
-            distro_like="debian",
-            distro_major_version="16",
-            distro_version="16.04",
-        ),
-        platform(
-            arch="x86_64",
-            os="Linux",
-            distro="gentoo",
-            distro_name="Gentoo",
-            distro_codename="n/a",
-            distro_major_version="2",
-            distro_version="2.7",
-        ),
-        platform(
-            arch="x86_64",
-            os="Darwin",
-            distro="darwin",
-            distro_name="Darwin",
-            distro_major_version="19",
-            distro_version="19.6.0",
-            mac_release="10.15.7",
-        ),
-        platform(
-            arch="AMD64",
-            os="Windows",
-            win_release="",
-        ),
-        platform(
-            arch="aarch64",
-            os="Linux",
-            # no distro in termux on android
-        ),
-    ),
-)
+@pytest.mark.parametrize("mock_platform", platform_samples)
 def test_mongo_not_running(rule_runner: RuleRunner, mock_platform: Platform) -> None:
     request = UsesMongoRequest(
         db_host="127.100.20.7",
